requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-identity
azure-core
python-dotenv
azure-search-documents 
azure-keyvault-secrets
azure-appconfiguration-provider
openai
azure-storage-blob
azure.ai.textanalytics